Movieproject/requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-decouple
drf-yasg
celery
redis
requests
django-filter
django-cors-headers